UML: Support User Mode Linux (UML) for certain test cases
The user has to set the environment variable IMA_TEST_UML to the full path
of the UML linux binary for UML-supporting test cases to run in UML.

UML test cases are setting up a chroot environment where only busybox is
available as well as the script files relevant for a test case. This
resembles the environment that is setup for container testing. However,
since devices such as /dev/null have to be created it is necessary to
run the UML tests as root.

Test cases for UML need to be converted to not use 'exit' but 'exit_test'
so that in a UML environment the test case can write its exit code into a
file (/__exitcode). The caller then reads the exit code from this file.

+++ ns-common.sh
@@ -11,6 +11,19 @@ EVM_SETUP_COMPLETE=$((0x80000000))
 
 # Common functions for the test scripts running in the namespace
 
+# Exit test wraps 'exit' but also allows to write the exit code into a
+# well known file on UML.
+exit_test()
+{
+  local val="$1"
+
+  if [ -n "${IMA_TEST_UML}" ]; then
+    echo "${val}" > /__exitcode
+  fi
+
+  exit "${val}"
+}
+
 # Mount the securityfs with IMA support; if it doesn't work or 
 # the ima directory doesn't show up exit with ${SKIP:-3}
 # @param1: mount directory
@@ -32,7 +45,7 @@ mnt_securityfs()
   local msg dmsgmsg timeout tmp
 
   # Skip mount if already mounted
-  tmp="$(mount -t securityfs | sed -n 's/^securityfs on \(.*\) type .*/\1/p')"
+  tmp="$(mount -t securityfs 2>/dev/null | sed -n 's/^securityfs on \(.*\) type .*/\1/p')"
   if [ "${tmp}" = "${mntdir}" ]; then
     return
   fi
@@ -82,7 +95,9 @@ mnt_securityfs()
     fi
   fi
 
-  echo 1 > "${mntdir}/ima/active"
+  if [ -f "${mntdir}/ima/active" ]; then
+    echo 1 > "${mntdir}/ima/active"
+  fi
 
   if [ -f "${mntdir}/integrity/evm/active" ]; then
     echo 1 > "${mntdir}/integrity/evm/active"
